startup/taskTune/requirements.txt
fastapi==0.105.0
uvicorn==0.24.0
sqlalchemy==2.0.23
pydantic==2.5.2
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.1.1
python-dotenv==1.0.0
psycopg2-binary==2.9.9
email-validator==2.0.0
openai==1.3.0
python-multipart==0.0.6
tenacity==8.2.3

